# environment-a/api-service/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AdminActivityPointDTO:
    properties:
      date:
        type: string
      deletes:
        type: integer
      downloads:
        type: integer
      logins:
        type: integer
      restores:
        type: integer
      stars:
        type: integer
      uploads:
        type: integer
    type: object
  dto.AdminAnalyticsResponse:
    properties:
      activity:
        items:
          $ref: '#/definitions/dto.AdminActivityPointDTO'
        type: array
      depths:
        items:
          $ref: '#/definitions/dto.AdminDepthStatDTO'
        type: array
      file_types:
        items:
          $ref: '#/definitions/dto.AdminFileTypeStatDTO'
        type: array
      generated_at:
        type: string
      range:
        type: string
      size_buckets:
        items:
          $ref: '#/definitions/dto.AdminSizeBucketStatDTO'
        type: array
      status:
        type: string
      summary:
        $ref: '#/definitions/dto.AdminAnalyticsSummaryDTO'
    type: object
  dto.AdminAnalyticsSummaryDTO:
    properties:
      active_files:
        type: integer
      active_folders:
        type: integer
      active_storage_bytes:
        type: integer
      active_users:
        type: integer
      deleted_items_in_range:
        type: integer
      downloads_in_range:
        type: integer
      restored_items_in_range:
        type: integer
      starred_actions_in_range:
        type: integer
      starred_files:
        type: integer
      starred_folders:
        type: integer
      total_admins:
        type: integer
      total_users:
        type: integer
      trash_files:
        type: integer
      trash_folders:
        type: integer
      trash_storage_bytes:
        type: integer
      uploads_in_range:
        type: integer
    type: object
  dto.AdminDepthStatDTO:
    properties:
      count:
        type: integer
      depth:
        type: integer
    type: object
  dto.AdminFileTypeStatDTO:
    properties:
      count:
        type: integer
      total_bytes:
        type: integer
      type:
        type: string
    type: object
  dto.AdminSizeBucketStatDTO:
    properties:
      bucket:
        type: string
      count:
        type: integer
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: unauthorized
        type: string
      status:
        example: error
        type: string
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      status:
        type: string
      user:
        $ref: '#/definitions/dto.UserDTO'
    type: object
...
  dto.MeResponse:
    properties:
      status:
        type: string
      user:
        $ref: '#/definitions/dto.UserDTO'
    type: object
  dto.OKResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      full_name:
        maxLength: 150
        minLength: 2
        type: string
      password:
        minLength: 8
        type: string
    required:
    - confirm_password
    - email
    - full_name
    - password
    type: object
...
  dto.UpdateProfileRequest:
    properties:
      current_password:
        type: string
      email:
        type: string
      full_name:
        maxLength: 150
        minLength: 2
        type: string
      new_password:
        minLength: 8
        type: string
    required:
    - email
    - full_name
    type: object
...
  dto.UserDTO:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      role:
        type: string
    type: object
info:
  contact: {}
  description: API service for authentication, directory hierarchy, and immutable
    file metadata.
  title: HashBox API Service
  version: "1.0"
paths:
...
  /admin/analytics:
    get:
      description: Return aggregate, privacy-safe application analytics for admin
        users
      parameters:
      - description: 'Analytics range: 7d, 30d, or 90d'
        in: query
        name: range
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AdminAnalyticsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Admin analytics
      tags:
      - admin
  /auth/login:
    post:
      consumes:
      - application/json
      description: Create bearer access token for the active session from email and
        password
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      description: Revoke current access session
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OKResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
  /auth/me:
    get:
      description: Return authenticated user profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Current user
      tags:
      - auth
    patch:
      consumes:
      - application/json
      description: Update authenticated user profile and optionally change password
      parameters:
      - description: Profile payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update profile
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with user role
      parameters:
      - description: Register payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Register user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"